init scripts: minor fixes

+++ src/scripts/init/pacman_colors.sh
@@ -3,6 +3,11 @@
 setPacmanColors() {
     local pacman_conf_f="/etc/pacman.conf"
     echo -e "\n\nEnabling Pacman colored output"
+
+    if [ ! -f "${pacman_conf_f}" ]; then
+        echo "${pacman_conf_f} not found, cannot continue"
+        return 1
+    fi
     cp -af "${pacman_conf_f}" "${pacman_conf_f}.bak"
     echo "Pacman config file backed up at ${pacman_conf_f}.bak"
     sed -i 's/#Color/Color\nILoveCandy/g' "${pacman_conf_f}"
